requirements.txt
Flask==2.3.2
flask-cors==5.0.0
openai==1.62.0
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3